requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
colorama==0.4.1
cryptography==2.5
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
huey==1.11.0
idna==2.8
names==0.3.0
netifaces==0.10.9
oauthlib==3.0.1
pycparser==2.19
pyotp==2.2.7
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
